Изменен тип столбцов под ресурсы

+++ database/migrations/2023_04_13_141756_create_orders_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->foreignIdFor(Consumer::class)->constrained()->cascadeOnDelete();
             $table->tinyInteger('resource')->default(Resources::ENERGY->value);
-            $table->unsignedBigInteger('resource_amount')->default(0);
+            $table->decimal('resource_amount', 22, 4)->default(0);
             $table->string('status')->default(Statuses::new->value);
             $table->timestamp('executed_at')->nullable();
             $table->timestamps();
